refactor main view into its component parts
