+++ desktop-exporter/app/main.css
@@ -7,6 +7,7 @@ body {
 
 .container {
   display: flex;
+  height: 100vh;
 }
 
 .detail {
